WinGui: - Fix for "del" key deleting items from the queue when additional modifier keys were pressed.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@3479 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ win/C#/frmQueue.cs
@@ -507,7 +507,7 @@ namespace Handbrake
         /// </param>
         private void ListQueueDeleteKey(object sender, KeyEventArgs e)
         {
-            if (e.KeyCode == Keys.Delete)
+            if (e.KeyCode == Keys.Delete && e.Modifiers == Keys.None)
                 DeleteSelectedItems();
         }
 
@@ -665,7 +665,7 @@ namespace Handbrake
         /// </param>
         private void MnuReaddClick(object sender, EventArgs e)
         {
-            if (!queue.LastEncode.IsEmpty)
+            if (queue.LastEncode != null && !queue.LastEncode.IsEmpty)
             {
                 queue.Add(
                     queue.LastEncode.Query,
